+++ test/integration/default/default_test.rb
@@ -17,3 +17,18 @@ end
 describe(command('cat /etc/passwd')) do
   its('stdout') { should match(/tomcat/) }
 end
+
+# validate that wget is installed
+describe package('wget') do
+  it { should be_installed }
+end
+
+# make sure that the download is successful
+describe file('/tmp/apache-tomcat-8.5.29.tar.gz') do
+  it { should exist }
+end
+
+# check for existance of tomcat server
+describe file('/opt/tomcat') do
+  it { should exist }
+end