+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # 2_Tomcat Cookbook Changelog
 
+# 0.2.0
+
+- Download Tomcat install
+
+  Deployed Tomcat tar file
+
 # 0.1.0
 
 - Initial development
+
   Deploy OpenJDK 7
+
   Deploy tomcat group
+
   Deploy tomcat user